requirements.txt
cffi
Cython
django
#django-admin-menu>=1.1
django-ldapdb
#GitPython
gunicorn
json2html
lxml
pyOpenSSL
requests
pyjnius
pytest
pytest-django
#psycopg2
psycopg2-binary
